Added some comments moved user_profile to components folder
I moved user_profile to the components folder since it isn't its own standalone screen at this point.
